refactor : 문자열 분리 로직을 StringSeparator로 이동
