+++ src/main/java/StringAddCalculator.java
@@ -1,11 +1,4 @@
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 public class StringAddCalculator {
-    private final static String COMMA = ",";
-    private final static String COLON = ":";
-    private final static String DELIMITER = ",|:";
-    private final static String CUSTOM_DELIMITER_PATTERN = "//(.)\\n(.*)";
 
     private final StringSeparator stringSeparator;
 
@@ -22,22 +15,7 @@ public class StringAddCalculator {
             return Integer.parseInt(input);
         }
 
-        if (containsDelimiter(input)) {
-            String[] numbers = input.split(DELIMITER);
-            return getSum(numbers);
-        }
-
-        Matcher m = Pattern.compile(CUSTOM_DELIMITER_PATTERN).matcher(input);
-        if (m.find()) {
-            String customDelimiter = m.group(1);
-            return getSum(m.group(2).split(customDelimiter));
-        }
-
-        return null;
-    }
-
-    private boolean containsDelimiter(String input) {
-        return input.contains(COMMA) || input.contains(COLON);
+        return getSum(stringSeparator.split(input));
     }
 
     private int getSum(String[] split) {
